Player's above average score
Python function aboveaverage(l) that takes a list of pairs of the form (name,score) as argument, where name is a string and score is an integer. Each pair is to be interpreted as the score of the named player. For instance, an input of the form [('Kohli',73),('Ashwin',33),('Kohli',7),('Pujara',122),('Ashwin',90)] represents two scores of 73 and 7 for Kohli, two scores of 33 and 90 for Ashwin and one score of 122 for Pujara. Your function should compute the list of players whose individual average score is greater than or equal to the overall average score. For an individual player, the average score is the total across all scores for that player divided by number of entries for that player. The overall average score is the total across all scores for all the players divided by the total number of entries across all players. The list should be sorted in ascending order by the name of the player.

For instance, aboveaverage([('Kohli',73),('Ashwin',33),('Kohli',7),('Pujara',122),('Ashwin',90)]) should return ['Pujara'] because the overall average score is 65 (325 divided by 5), Kohli's average is 40, (80 divided by 2), Ashwin's average is 61.5 (123 divided by 2) and Pujara's average is 122 (122 divided by 1).